Rename database utilities.
