+++ estimagic/optimization/optimize.py
@@ -7,11 +7,11 @@ import numpy as np
 import estimagic.batch_evaluators as be
 from estimagic.config import ALL_ALGORITHMS
 from estimagic.config import DEFAULT_DATABASE_NAME
-from estimagic.logging.database_utilities_new import append_row
-from estimagic.logging.database_utilities_new import load_database
-from estimagic.logging.database_utilities_new import make_optimization_iteration_table
-from estimagic.logging.database_utilities_new import make_optimization_problem_table
-from estimagic.logging.database_utilities_new import make_optimization_status_table
+from estimagic.logging.database_utilities import append_row
+from estimagic.logging.database_utilities import load_database
+from estimagic.logging.database_utilities import make_optimization_iteration_table
+from estimagic.logging.database_utilities import make_optimization_problem_table
+from estimagic.logging.database_utilities import make_optimization_status_table
 from estimagic.optimization.broadcast_arguments import broadcast_arguments
 from estimagic.optimization.check_arguments import check_argument
 from estimagic.optimization.internal_criterion_template import (
